fix: proibindo exclusão de projetos com tarefas atribuídas

+++ src/http/controller/projects/deleteProject.ts
@@ -10,11 +10,27 @@ export async function deleteProject(request: FastifyRequest, reply: FastifyReply
         return reply.status(403).send({ message: 'Acesso negado.' })
     }
 
-    const project = await prisma.project.delete({
+    const project = await prisma.project.findUnique({
+    where: {
+      id: Number(targetProjectId),
+    },
+    include: {
+      tasks: true,
+    },
+  })
+    if (!project) {
+        return reply.status(404).send({ message: 'Projeto não encontrado.' })
+    }
+    if (project.tasks.length > 0) {
+        return reply.status(400).send({ message: 'Não é possível deletar um projeto que possui tarefas associadas.' })
+    }
+    
+    prisma.project.delete({
         where: {
             id: Number(targetProjectId),
         },
     })
+
     const projectResponse = {
         name: project.name,
         description: project.description,
